MSVC still broken and bug fix

+++ groups/bsl/bslfmt/bslfmt_format.t.cpp
@@ -868,7 +868,7 @@ struct IntFormatter {
         ++g_numFormatInvocations;
         typename t_FORMAT_CONTEXT::iterator outIterator = context.out();
         g_values.push_back(value.value());
-        int divider = sizeof(int) == 4 ? 1000000000 : 1000000000000000000;
+        long divider = sizeof(int) == 4 ? 1000000000 : 1000000000000000000;
             // INT_MAX 9,223,372,036,854,775,807 on 64-bit INT_MAX
             // 2,147,483,647 on 32-bit
